renamed some things and tried to make everything consistent
