+++ src/main/java/com/revature/ecommerce_cli/Models/OrderProduct.java
@@ -1,5 +1,4 @@
-package com.revature.ecommerce_cli.models;
-
+package com.revature.ecommerce_cli.Models;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,19 +6,19 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Setter
 @ToString
 
-public class Users {
-    
+
+public class OrderProduct {
+
     private String id;
-    private String username;
-    private String password;
+    private String order_id;
+    private String product_id;
+    private int quantity; 
     
-
-    
-
 }
