+++ src/main/java/com/revature/ecommerce_cli/Models/Review.java
@@ -1,4 +1,5 @@
-package com.revature.ecommerce_cli.models;
+package com.revature.ecommerce_cli.Models;
+
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,19 +7,18 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
-
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Setter
 @ToString
 
-
-public class OrderProducts {
+public class Review {
 
     private String id;
-    private String order_id;
+    private int rating;
+    private String comment;
+    private String user_id;
     private String product_id;
-    private int quantity; 
-    
+
 }
